+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Changed
 
 - `delete bulk`: slight performance optimisations.
+- `create dataset`: Accept an optional `--model-family` and `--copy-annotations-from` for the new dataset.
 
 # v0.5.3
 
@@ -36,7 +37,6 @@ re create dataset org/example-dataset -s org/example-source --has-sentiment fals
 
 ## Changed
 
-- `create dataset`: Accept an optional `--model-family` and `--copy-labels-from` for the new dataset.
 - `create bucket`: Accept an optional `--transform-tag` value for the new bucket.
 - `get buckets`: Display transform tag for retrieved data.